Simple Limit Counter: polish weird failure description
An answer to a QuickQuiz about the 'Simple Limit Counter' explains about
weird failure of sub_count().  It says the possible failure is weird
because the logical count is non-zero in actual.  However, because the
interfaces do not restrict users to add and/or subtract 1 only, the
failure can be normal if the logical value was larger than zero and
argument to sub_count() was even larger than that.  So, the sentence
could be not accurate enough to some reader.  This commit polish the
sentence to explots the meaning better by adapting words from 'Simple
Limit Counter Discussion' section.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>